AKCORE-129-2: Bugfix for the issue where partition maxInFlightMessages limit was getting exceeded with multiple consumers (#1224)
With the previous code, when there were multiple consumers consuming from a single topic partition with a huge number of records, maxInFlightMessages limit was getting exceeded. This was because acknowledging and fetching could happen simultaneously, both of which can change the nextFetchOffset, potentially causing inconsistencies.

With the new changes in place, the variable nextFetchOffset has been removed entirely. The logic in nextFetchOffset method has been changed.

The new logic to find the next offset to start fetching from depends on findNextFetchOffset flag. If this flag is set to true, then the next fetch offset should be recomputed. If the flag is false, then the next fetch offset is Share Partition End Offset + 1. The flag is set to true in the following cases -

When some previously acquired records are acknowledged with type RELEASE
When the record lock duration expires for some acquired records using the method releaseAcquisitionLockOnTimeout
When some records are released on final Fetch/ Acknowledge request using the method releaseAcquiredRecords
The re-computation of next fetch offset is done by iterating over the cachedState and finding the first available record. If no available record is found, then the next fetch offset is set to Share Partition End Offset + 1 and findNextFetchOffset flag is set to false